[Feat] qna 부분 엔티티 관계키 주석 해제 및 연결
