+++ backend/src/main/java/org/example/backend/Answer/Model/Entity/AnswerLike.java
@@ -1,10 +1,11 @@
-package org.example.backend.Answer.Entity;
+package org.example.backend.Answer.Model.Entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.example.backend.User.Model.Entity.User;
 
 
 @Entity
@@ -17,12 +18,12 @@ public class AnswerLike {
         @GeneratedValue(strategy = GenerationType.IDENTITY)
         private Long id; // bigint
 
-//        @ManyToOne(fetch = FetchType.LAZY)
-//        @Column(name = "user_id", nullable = false)
-//        private User user;
+        @ManyToOne(fetch = FetchType.LAZY)
+        @JoinColumn(name = "user_id", nullable = false)
+        private User user;
 
         @ManyToOne(fetch = FetchType.LAZY)
-        @Column(name = "answer_id", nullable = false)
+        @JoinColumn(name = "answer_id", nullable = false)
         private Answer anser;
 
         @Column(name = "state", nullable = false)
