+++ backend/src/main/java/org/example/backend/Qna/model/Entity/QnaLike.java
@@ -1,10 +1,11 @@
-package org.example.backend.Qna.Entity;
+package org.example.backend.Qna.model.Entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.example.backend.User.Model.Entity.User;
 
 
 @Entity
@@ -17,9 +18,9 @@ public class QnaLike {
         @GeneratedValue(strategy = GenerationType.IDENTITY)
         private Long id;
 
-//        @ManyToOne(fetch = FetchType.LAZY)
-//        @Column(name = "user_id", nullable = false)
-//        private User user;
+        @ManyToOne(fetch = FetchType.LAZY)
+        @JoinColumn(name = "user_id", nullable = false)
+        private User user;
 
         @ManyToOne(fetch = FetchType.LAZY)
         @JoinColumn(name = "qna_board_id", nullable = false)
